[docs] file명 수정, preview 기능 구현중
구현과정중 문제가 생겨 모달만들기 먼저 진행
